+++ astroplan/tests/test_observer.py
@@ -103,7 +103,7 @@ def test_altaz_multiple_targets():
     times = Time('1995-06-21 00:00:00') + np.linspace(0, 1, 5)*u.day
 
     obs = Observer(location=location)
-    transformed_coords = obs.altaz(times, targets)
+    transformed_coords = obs.altaz(times, targets, grid_times_targets=True)
     altitudes = transformed_coords.alt
 
     # Double check by doing one star the normal way with astropy
@@ -121,7 +121,7 @@ def test_altaz_multiple_targets():
 
     # check that output elements are the proper lengths and types
     assert isinstance(vega_list_alt, Latitude)
-    assert len(vega_list_alt[0, :]) == len(times)
+    assert len(vega_list_alt) == len(times)
 
     # Check for single time
     single_time = times[0]
@@ -142,7 +142,7 @@ def test_altaz_multiple_targets():
     capella_FixedTarget = FixedTarget(coord=capella, name='Capella')
     sirius_FixedTarget = FixedTarget(coord=sirius, name='Sirius')
     ft_list = [vega_FixedTarget, capella_FixedTarget, sirius_FixedTarget]
-    ft_vector_alt = obs.altaz(times, ft_list).alt
+    ft_vector_alt = obs.altaz(times[:, np.newaxis], ft_list).T.alt
     assert all(ft_vector_alt[0, :] == vega_alt)
     assert all(ft_vector_alt[2, :] == sirius_alt)
 
@@ -1016,7 +1016,7 @@ def test_TargetNeverUpWarning(recwarn):
     assert no_time == MAGIC_TIME
 
 
-def test_mixed_rise_and_dont_rise(recwarn):
+def test_mixed_rise_and_dont_rise():
     vega = SkyCoord(279.23473479*u.deg, 38.78368896*u.deg)
     polaris = SkyCoord(37.95456067*u.deg, 89.26410897*u.deg)
     sirius = SkyCoord(101.28715533*u.deg, -16.71611586*u.deg)
@@ -1026,7 +1026,8 @@ def test_mixed_rise_and_dont_rise(recwarn):
     time = Time('1995-06-21 00:00:00')
 
     obs = Observer(location=location)
-    rise_times = obs.target_rise_time(time, targets, which='next')
+    with pytest.warns(TargetAlwaysUpWarning) as recwarn:
+        rise_times = obs.target_rise_time(time, targets, which='next')
 
     assert rise_times[1] == MAGIC_TIME
 
@@ -1207,15 +1208,15 @@ def test_tonight():
 
     during_day = obs.tonight(time=noon, horizon=horizon)
     during_night = obs.tonight(time=midnight, horizon=horizon)
-    
-    assert (abs(sunset - during_day[0].datetime) < 
+
+    assert (abs(sunset - during_day[0].datetime) <
         datetime.timedelta(minutes=threshold_minutes))
-    assert (abs(sunrise - during_day[1].datetime) < 
+    assert (abs(sunrise - during_day[1].datetime) <
         datetime.timedelta(minutes=threshold_minutes))
 
-    assert (abs(midnight.datetime - during_night[0].datetime) < 
+    assert (abs(midnight.datetime - during_night[0].datetime) <
         datetime.timedelta(minutes=threshold_minutes))
-    assert (abs(sunrise - during_night[1].datetime) < 
+    assert (abs(sunrise - during_night[1].datetime) <
         datetime.timedelta(minutes=threshold_minutes))
 
     astro_sunset = Time('2016-08-08 06:13:00')