+++ CHANGES.rst
@@ -1,7 +1,12 @@
 0.3 (unreleased)
 ----------------
 
-- No changes yet
+- ``Observer.altaz`` and ``Constraint.__call__`` no longer returns an (MxN) grid of results when
+  called with M ``target``s and N ``times``. Instead, we attempt to broadcast the time and target
+  shapes, and an error is raised if this is not possible. This change breaks backwards compatibility
+  but an optional argument ``grid_times_targets`` has been added to these methods. If set to true,
+  the old behaviour is recovered. All ``Observer`` methods for which it is relevant have this optional
+  argument.
 
 0.2.1 (2016-04-27)
 ------------------
@@ -11,8 +16,9 @@
 - Fixed bug when scheduling block list is empty [#298]
 
 - Fixed bug in Transitioner object when no transition needed [#295]
- 
+
 - Update to astropy-helpers 1.3.1 [#294] and compatibility fixes for astropy 1.3 [#283]
+
 
 0.2 (2016-09-20)
 ----------------